requirements.txt
discord.py>=2.0.0
google-api-python-client>=2.39.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
tabulate>=0.8.9